requirements.txt
# How to install: pip3 install -r requirements.txt
Scrapy==2.5.0
SQLAlchemy==1.4.25
jsonpickle==2.0.0
requests==2.26.0
pymongo==3.12.0
htmlmin==0.1.12
Flask_SQLAlchemy==2.5.1
itemadapter==0.4.0
Flask==2.0.1
Flask-MySQL==1.5.2
dnspython == 2.1.0
requests==2.26.0
beautifulsoup4==4.10.0
scrapy-rotating-proxies==0.6.2
mysql-connector-python==8.0.26
sauth==1.0.1
loguru
PySocks
openpyxl
openpyxl-templates
future
scrapy-zyte-smartproxy
python-dotenv